+++ examples/webhook_versioning_app.py
@@ -88,4 +88,4 @@ versioner.versionize()
 if __name__ == "__main__":
     import uvicorn
 
-    uvicorn.run(app, host="127.0.0.1", port=8004)
+    uvicorn.run(app, host="127.0.0.1", port=8006)